Fixed some problems and changed stuff around again. added a friction variable. implimented friction in an ugly way. new player testing sprite

+++ src/net/KoopaECS/Components/Transform.java
@@ -6,8 +6,8 @@ public class Transform {
 	public double x;
 	public double y;
 	
-	public int directionX;
-	public int directionY;
+	public double directionX;
+	public double directionY;
 	
 	
 	public double angle;
